+++ uploader/utils/languages.py
@@ -1,0 +1,55 @@
+languages = [
+    {"name": "English", "two_letter": "en", "three_letter": "eng"},
+    {"name": "Japanese", "two_letter": "ja", "three_letter": "jpn"},
+    {"name": "Japanese (Romaji)", "two_letter": "ja-ro", "three_letter": "jpn"},
+    {"name": "Polish", "two_letter": "pl", "three_letter": "pol"},
+    {"name": "Serbo-Croatian", "two_letter": "sh", "three_letter": "hrv"},
+    {"name": "Dutch", "two_letter": "nl", "three_letter": "dut"},
+    {"name": "Italian", "two_letter": "it", "three_letter": "ita"},
+    {"name": "Russian", "two_letter": "ru", "three_letter": "rus"},
+    {"name": "German", "two_letter": "de", "three_letter": "ger"},
+    {"name": "Hungarian", "two_letter": "hu", "three_letter": "hun"},
+    {"name": "French", "two_letter": "fr", "three_letter": "fre"},
+    {"name": "Finnish", "two_letter": "fi", "three_letter": "fin"},
+    {"name": "Vietnamese", "two_letter": "vi", "three_letter": "vie"},
+    {"name": "Greek", "two_letter": "el", "three_letter": "gre"},
+    {"name": "Bulgarian", "two_letter": "bg", "three_letter": "bul"},
+    {"name": "Spanish (Es)", "two_letter": "es", "three_letter": "spa"},
+    {"name": "Portuguese (Br)", "two_letter": "pt-br", "three_letter": "por"},
+    {"name": "Portuguese (Pt)", "two_letter": "pt", "three_letter": "por"},
+    {"name": "Swedish", "two_letter": "sv", "three_letter": "swe"},
+    {"name": "Arabic", "two_letter": "ar", "three_letter": "ara"},
+    {"name": "Danish", "two_letter": "da", "three_letter": "dan"},
+    {"name": "Chinese (Simp)", "two_letter": "zh", "three_letter": "chi"},
+    {"name": "Chinese (Romaji)", "two_letter": "zh-ro", "three_letter": "chi"},
+    {"name": "Bengali", "two_letter": "bn", "three_letter": "ben"},
+    {"name": "Romanian", "two_letter": "ro", "three_letter": "rum"},
+    {"name": "Czech", "two_letter": "cs", "three_letter": "cze"},
+    {"name": "Mongolian", "two_letter": "mn", "three_letter": "mon"},
+    {"name": "Turkish", "two_letter": "tr", "three_letter": "tur"},
+    {"name": "Indonesian", "two_letter": "id", "three_letter": "ind"},
+    {"name": "Korean", "two_letter": "ko", "three_letter": "kor"},
+    {"name": "Korean (Romaji)", "two_letter": "ko-ro", "three_letter": "kor"},
+    {"name": "Spanish (LATAM)", "two_letter": "es-la", "three_letter": "spa"},
+    {"name": "Persian", "two_letter": "fa", "three_letter": "per"},
+    {"name": "Malay", "two_letter": "ms", "three_letter": "may"},
+    {"name": "Thai", "two_letter": "th", "three_letter": "tha"},
+    {"name": "Catalan", "two_letter": "ca", "three_letter": "cat"},
+    {"name": "Filipino", "two_letter": "tl", "three_letter": "fil"},
+    {"name": "Chinese (Trad)", "two_letter": "zh-hk", "three_letter": "chi"},
+    {"name": "Ukrainian", "two_letter": "uk", "three_letter": "ukr"},
+    {"name": "Burmese", "two_letter": "my", "three_letter": "bur"},
+    {"name": "Lithuanian", "two_letter": "lt", "three_letter": "lit"},
+    {"name": "Hebrew", "two_letter": "he", "three_letter": "heb"},
+    {"name": "Hindi", "two_letter": "hi", "three_letter": "hin"},
+    {"name": "Norwegian", "two_letter": "no", "three_letter": "nor"},
+    {"name": "Latin", "two_letter": "la", "three_letter": "lat"},
+    {"name": "Kazakh", "two_letter": "kk", "three_letter": "kaz"},
+    {"name": "Tamil", "two_letter": "ta", "three_letter": "tam"},
+    {"name": "Croatian", "two_letter": "hr", "three_letter": "hrv"},
+    {"name": "Esperanto", "two_letter": "eo", "three_letter": "epo"},
+    {"name": "Estonian", "two_letter": "et", "three_letter": "est"},
+    {"name": "Nepali", "two_letter": "ne", "three_letter": "nep"},
+    {"name": "Serbian", "two_letter": "sr", "three_letter": "srp"},
+    {"name": "Other", "two_letter": "null", "three_letter": "null"},
+]
